+++ app/src/core/providers/adapters/ollama/index.ts
@@ -1,0 +1,2 @@
+export * from './ollamaTypes'
+export * from './ollamaAdapter'